Mid work on email functionality
