+++ wap
@@ -8,21 +8,21 @@ from datetime import datetime, timezone
 from describe import describe
 from approve import auto_approve
 from session import Session
-from discourse import Discourse, sync_group
+#from discourse import Discourse, sync_group
 import sys
 from tzlocal import get_localzone
 
 now = datetime.now(timezone.utc)
 session = Session()
 contacts = Contacts(session)
-dc_client = Discourse(contacts)
+#dc_client = Discourse(contacts)
 
 operations = {
         'archive': lambda x: archive(contacts, now),
         'describe': lambda x: describe(contacts, x),
         'cancel': lambda x: autoCancel(contacts),
-        'approve' : lambda x: auto_approve(session),
-        'discourse_sync': lambda x: sync_group(dc_client, x)}
+        'approve' : lambda x: auto_approve(session)}
+        #'discourse_sync': lambda x: sync_group(dc_client, x)}
 
 for op in args.operations:
     if op[0] not in operations:
